sp1 to sp1

See merge request chengdu/deepin-album!14

+++ widgets/pushbutton.cpp
@@ -214,10 +214,9 @@ void PushButton::paintEvent(QPaintEvent *e)
             const QRect pr(m.left(), (height() - ph) / 2, ph, ph);
             painter.drawPixmap(QPoint(pr.x(), pr.y()), pixmap
                                /*pixmap.scaled(pr.size(), Qt::KeepAspectRatioByExpanding)*/);
-        }
-        else {
-            ph = pixHeight;
-            const QRect pr(m.left(), (height() - ph) / 2, pixWidth, ph);
+        } else {
+            ph = static_cast<int>(pixHeight);
+            const QRect pr(m.left(), (height() - ph) / 2, static_cast<int>(pixWidth), ph);
             painter.drawPixmap(QPoint(pr.x(), pr.y()), pixmap);
         }
         spacing = m_spacing;
@@ -227,7 +226,7 @@ void PushButton::paintEvent(QPaintEvent *e)
     const int tw = width() - m.left() - spacing - ph - m.right();
     const int th = fm.height();
     const QRect textRect(m.left() + ph + spacing, (height() - th) / 2,
-                   tw, th);
+                         tw, th);
     const QString mt = fm.elidedText(m_text, Qt::ElideMiddle, tw);
     painter.setPen(QPen(getTextColor()));
     painter.drawText(textRect, Qt::AlignCenter, mt);
@@ -292,11 +291,9 @@ QString PushButton::getPixmap() const
 {
     if (m_checked) {
         return checkedPic();
-    }
-    else if (isEnabled()) {
+    } else if (isEnabled()) {
         return m_currentPic;
-    }
-    else {
+    } else {
         return disablePic();
     }
 }
@@ -305,8 +302,7 @@ QColor PushButton::getTextColor() const
 {
     if (isEnabled()) {
         return m_currentColor;
-    }
-    else {
+    } else {
         return disableColor();
     }
 }
@@ -330,7 +326,7 @@ void PushButton::showTooltip(const QPoint &pos)
     if (y > dr.y() + dr.height()) {
         y = pos.y() - tf->height() - 10;
     }
-    tf->move(pos.x() - tf->width()/3, y - tf->height()/3);
+    tf->move(pos.x() - tf->width() / 3, y - tf->height() / 3);
 
     QTimer::singleShot(5000, tf, SLOT(deleteLater()));
