+++ widgets/returnbutton.cpp
@@ -218,14 +218,14 @@ void ReturnButton::paintEvent(QPaintEvent *e)
             painter.drawPixmap(QPoint(pr.x(), pr.y()), pixmap
                                /*pixmap.scaled(pr.size(), Qt::KeepAspectRatioByExpanding)*/);
         } else {
-            ph = pixHeight;
-            const QRect pr(m.left(), (height() - ph) / 2, pixWidth, ph);
+            ph = static_cast<int>(pixHeight);
+            const QRect pr(m.left(), (height() - ph) / 2, static_cast<int>(pixWidth), ph);
             painter.drawPixmap(QPoint(pr.x(), pr.y()), pixmap);
         }
     }
 
     QFontMetrics fm(font());
-    int maxWidth = m_maxWidth - pixWidth - 6;
+    int maxWidth = static_cast<int>(m_maxWidth - pixWidth - 6);
     int textWidth = fm.boundingRect(m_text).width();
     QString mt;
     if (textWidth > maxWidth) {
@@ -234,7 +234,7 @@ void ReturnButton::paintEvent(QPaintEvent *e)
         mt = m_text;
     }
     textWidth = fm.boundingRect(mt).width();
-    setFixedWidth(textWidth + pixWidth + 6);
+    setFixedWidth(static_cast<int>(textWidth + pixWidth + 6));
 
     int oldWidth = m_buttonWidth;
     m_buttonWidth = std::max(24, int(textWidth + pixWidth + 6));
@@ -242,7 +242,7 @@ void ReturnButton::paintEvent(QPaintEvent *e)
         emit returnBtnWidthChanged(m_buttonWidth);
     }
     const int th = fm.height();
-    QRect textRect = QRect(pixWidth, (height() - th) / 2 - 1, textWidth, pixHeight);
+    QRect textRect = QRect(static_cast<int>(pixWidth), (height() - th) / 2 - 1, textWidth, static_cast<int>(pixHeight));
     painter.setPen(QPen(getTextColor()));
     painter.drawText(textRect, Qt::AlignCenter, mt);
     QWidget::paintEvent(e);
